Adds Pundit authorization

+++ app/helpers/application_helper.rb
@@ -15,6 +15,15 @@ module ApplicationHelper
 
   def destroy_link_for(object)
     return unless object.persisted?
+    return unless policy(object).destroy?
+
     link_to 'Delete', object, method: :delete, data: { confirm: 'Are you sure?' }, class: 'btn btn-danger'
+  end
+
+  def edit_link_for(object)
+    return unless object.persisted?
+    return unless policy(object).edit?
+
+    link_to 'Edit', polymorphic_path([:edit, object]), class: 'btn btn-info'
   end
 end
